requirements.txt
Django==4.2.10
djangorestframework==3.14.0
drf_yasg==1.21.5
